[Retry] Edit 페이지 구현

+++ src/pages/Edit.jsx
@@ -2,7 +2,7 @@ import { useParams, useNavigate } from "react-router-dom";
 import Header from "./../components/Header";
 import Button from "../components/Button";
 import Editor from "../components/Editor";
-import { useContext } from "react";
+import { useContext, useEffect, useState } from "react";
 import { DiaryDispatchContext, DiaryStateContext } from "./../App";
 
 const Edit = () => {
@@ -10,15 +10,37 @@ const Edit = () => {
   const nav = useNavigate();
   const data = useContext(DiaryStateContext);
   const { onDelete, onUpdate } = useContext(DiaryDispatchContext);
+  const [curDiaryItem, setCurDiaryItem] = useState();
+
+  useEffect(() => {
+    const currentDiarayItem = data.find(
+      (item) => Number(item.id) === Number(params.id)
+    );
+    if (!currentDiarayItem) {
+      window.alert("존재하지 않는 일기입니다.");
+      nav("/", { replace: true });
+      return;
+    }
+    setCurDiaryItem(currentDiarayItem);
+  }, [params.id]);
 
   const onClick = () => {
-    onDelete(params.id);
-    nav("/", { replace: true });
+    if (window.confirm("삭제 하시겠습니까?")) {
+      onDelete(params.id);
+      nav("/", { replace: true });
+    }
   };
 
   const onSubmit = (input) => {
-    onUpdate(params.id, input.createdDate, input.emotionId, input.content);
-    nav("/", { replace: true });
+    if (window.confirm("수정 하시겠습니까?")) {
+      onUpdate(
+        params.id,
+        input.createdDate.getTime(),
+        input.emotionId,
+        input.content
+      );
+      nav("/", { replace: true });
+    }
   };
 
   return (
@@ -30,10 +52,7 @@ const Edit = () => {
           <Button text={"삭제하기"} type={"NEGATIVE"} onClick={onClick} />
         }
       />
-      <Editor
-        onSubmit={onSubmit}
-        diary={data.filter((item) => Number(item.id) === Number(params.id))}
-      />
+      <Editor onSubmit={onSubmit} initData={curDiaryItem} />
     </div>
   );
 };
